+++ lib/NonlinearSolveBase/ext/NonlinearSolveBaseSparseArraysExt.jl
@@ -15,6 +15,9 @@ Utils.make_sparse(x) = sparse(x)
 
 Utils.condition_number(J::AbstractSparseMatrix) = Utils.condition_number(Matrix(J))
 
-Utils.maybe_pinv!!_workspace(A::AbstractSparseMatrix) = Matrix(A)
+function Utils.maybe_pinv!!_workspace(A::AbstractSparseMatrix)
+    dense_A = Matrix(A)
+    return dense_A, copy(dense_A)
+end
 
 end